refactor: useSnackbar에 message 와 컴포넌트 추가
